+++ CHANGELOG.md
@@ -2,14 +2,23 @@
 
 ## develop (unreleased)
 
+### New features
+
 * New defcustom `orgbox-start-of-day` allows users to customize the
   start time of day.
+* New defcustom `orgbox-start-of-weekends` allows users to customize
+  the start time of weekends.
 * New defcustom `orgbox-start-of-evening` allows users to customize
   the start time of evening.
 * New defcustom `orgbox-later` allows users to customize 'later'.
 * New defcustom `orgbox-someday` allows users to customize 'someday'.
+
+### Changes
+
 * Replaced function `orgbox-later-today` with
   `orgbox-schedule-later-today`.
+* Replaced function `orgbox-this-or-next-weekend` with
+  `orgbox-schedule-this-or-next-weekend`.
 * Replaced function `orgbox-this-or-tomorrow-evening` with
   `orgbox-schedule-this-or-tomorrow-evening`.
 * Replaced function `orgbox-tomorrow` with `orgbox-schedule-tomorrow`.